Simple 3D Music Visualizer
This is a Music Vizualizer created in Processing, a software sketchbook that uses Java as a foundation.

What this code does is that it projects a series of shapes moving in a 3D space. The shapes change in size and opacity based off of the volume of the soundtrack, and the coordinates of the mouse on the canvas dictate the color values.

The program also comes with microphone integration--additional sound input will contribute to the Blue values of the shapes ('0' being totally silent and '255' being maximum volume). They also temporarily increase the size of the shapes.